requirements.txt
beautifultable==1.0.1
filelock==3.0.12
fire==0.4.0
pyaes==1.6.1
pyasn1==0.4.8
pysondb==1.3.0
python-dotenv==0.19.0
rsa==4.7.2
six==1.16.0
Telethon==1.23.0
termcolor==1.1.0
wcwidth==0.2.5
